Refactor project structure for readability and maintainability
This change reorganizes the project to follow common Python conventions, making it easier for new developers to understand and contribute to.

Key changes:
- Created a main `app` directory to house the application and its source code.
- Moved the main `app.py` script and the `src` directory into the `app` directory.
- Turned the `app` directory into a Python package by adding an `__init__.py` file.
- Created a `scripts` directory for the run command.
- Created a `tests` directory as a placeholder for future tests.
- Removed the `sys.path` manipulation from `app.py` and updated the imports to be absolute from the project root.
