+++ app/src/curriculum.py
