+++ app/src/panels/registrar_panel.py
