+++ app/src/utils/desktop.ini
