+++ app/src/utils/registrar_main_report_helper.py
