+++ app/src/utils/registration_helper.py
