+++ scripts/runapp.cmd
@@ -1,0 +1,1 @@
+streamlit run app/app.py
